requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.173
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
psycopg2==2.9.6
pydantic==1.10.7
python-environ==0.4.54
PyYAML==6.0
requests==2.30.0
SQLAlchemy==2.0.14
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
